+++ Sources/swiftl10n/EntryPoint.swift
@@ -5,7 +5,7 @@ struct SwiftL10nCommand: ParsableCommand {
     static let configuration = CommandConfiguration(
         commandName: "swiftl10n",
         abstract: "Detect hardcoded localizable strings in Swift projects and generate a strongly-typed i18n API.",
-        version: "0.5.1",
+        version: "0.5.2",
         subcommands: [ScanCommand.self, InitCommand.self],
         defaultSubcommand: ScanCommand.self
     )